+++ src/lexic/mod.rs
@@ -1,0 +1,11 @@
+// Modules de gestion du dictionnaire
+
+
+mod lx_portail;
+pub use self::lx_portail::Portail;
+
+
+mod lx_application;
+
+mod lx_utils;
+