+++ src/routic/mod.rs
@@ -33,6 +33,10 @@
 /// /eddy/document/:key
 /// /eddy/log
 
+mod rx_login;
+pub use self::rx_login::login; // curl http://0.0.0.0:8080/login
+pub use self::rx_login::logout; // curl http://0.0.0.0:8080/logout
+
 mod rx_portail;
 pub use self::rx_portail::portail; // curl http://0.0.0.0:8080/