+++ src/routic/rx_portail.rs
@@ -4,15 +4,17 @@ use actix_web::{
     // post,
     // HttpResponse,
     web,
+    web::ReqData,
     Responder,
     Result,
 };
 // use log::info;
 use askama::Template;
 use actix_web_lab::respond::Html;
+use actix_session::Session;
 
 use crate::AppState;
-
+use crate::servic;
 
 
 #[derive(Template)]
@@ -23,10 +25,17 @@ struct PortailTemplate {
     applications: Vec<String>,
 }
 
-// http://127.0.0.1:8080/portail
+// cuerl http://0.0.0.0:8080/
 #[get("/")]
-pub async fn portail(data: web::Data<AppState>) -> Result<impl Responder> {
-    // let portail = dx::Portail::new();
+pub async fn portail(session: Session, data: web::Data<AppState>, msg: Option<ReqData<servic::sx_data::Msg>>) -> Result<impl Responder> {
+    log::info!("Session {:?} {:?}", session.status(), session.entries());
+
+    if let Some(msg_data) = msg {
+        let servic::sx_data::Msg(message) = msg_data.into_inner();
+        log::info!("Msg: {:?}", message);
+    } else {
+        log::error!("no message found");
+    }
 
     let html = PortailTemplate {
         title: data.portail.title.clone(),