+++ src/servic/mod.rs
@@ -1,0 +1,6 @@
+
+pub mod sx_data;
+
+// use self::servic_data::AddMsg;
+
+pub mod sx_redirect;